Updating Compass

+++ lib/compass/sass_extensions/sprites/sprite_methods.rb
@@ -68,7 +68,8 @@ module Compass
         end
         
         def cleanup_old_sprites
-          Dir[File.join(Compass.configuration.images_path, "#{path}-s*.png")].each do |file|
+          Dir[File.join(Compass.configuration.generated_images_path, "#{path}-s*.png")].each do |file|
+            next if file[filename]
             log :remove, file
             FileUtils.rm file
             Compass.configuration.run_sprite_removed(file)
